only take affect in correct server

+++ index.js
@@ -807,12 +807,12 @@ client.on("message", async (message) => {
 	// handle banned words
 	const bannedWords = ["fuck", "nigg", "fuk", "cunt", "cnut", "bitch", "dick", "d1ck", "pussy", "asshole", "b1tch", "b!tch", "blowjob", "cock", "c0ck", "retard", "fag", "faggot"]
 	
-	if (bannedWords.some(word => message.content.toLowerCase().includes(word)) && !client.hasModPerms(client, message.member)) {
+	if (bannedWords.some(word => message.content.toLowerCase().includes(word)) && !client.hasModPerms(client, message.member) && message.guild.id === client.config.mainServer.id) {
 	message.delete()
 	message.reply("That word is banned here.").then(x => setTimeout(() => x.delete(), 5000))}
 
 	// handle discord invite links
-	if (message.content.includes('discord.gg/') && (!message.member.roles.cache.has(client.config.mainServer.roles.levels.three.id))) {
+	if (message.content.includes('discord.gg/') && (!message.member.roles.cache.has(client.config.mainServer.roles.levels.three.id)) && message.guild.id === client.config.mainServer.id) {
 		message.delete()
 		client.punishments.addPunishment('warn', message.member, { reason: 'Discord advertisement' }, client.user.id)
 		message.reply('No advertising other Discord servers.').then(x => setTimeout(() => x.delete(), 10000))
